MGQD & SGQD - new vars, edit func getFlux
New variables are to store 2D fluxes and currents of the previous
time step, which are used in the group collapse calculation.

getFlux now pulls in the interface fluxes and currents, in addition
to the cell average fluxes.

MGQD now has a getFlux function, as opposed to having that
functionality tacked on after the solve call. This is necessary,
as the getFlux function now needs the fluxes and currents to both
be solved for before being called.