MG->GG - split a func, new funcs, calcZetas
I changed the collapseNuclearData function to call a bunch of
newer smaller functions to encapsulate things better. Also added
in logic for calculating the zeta terms that result from collapsing
data using the absolute values of the currents.

Still need to collapse the outward current to flux ratios in the
calculateFluxWeightedData function.