MG->GG - new funcs, edit collapse funcs
Added a rough function to iterate between the MGQD and GGQD stages
to solve for a single time step.

Also edited the collapse functions to determine the necessity of
and implement a biasing factor when performing the flux weighted
data to prevent division by zero.

Had to add some new functions to check for zero BC fluxes across
all groups.