MLCoupling - iteration control knobs
Added maxResidual and overSolveThreshold parameters and appropriate
checks. On an ELOT solve, if the max residual is exceeded, jump to
the MGLOQD level. If the overSolveThreshold is exceeded, jump to
the MGLOQD level. These two knobs allow for an optimal number of
iterates to be achieved.

Also changed residual measurements to look at the residuals in
each different physics group.