Fixed PETSc bugs that were causing memory leaks
There were three types of issues common to the modified class
functions.

First, a variable holding broadcasted values does not need to be
initialized before the broadcast. Second, a global vector should
be destroyed before having values broadcast into it, if it is
subject to multiple broadcasts. Lastly, and most obviously, a solver
must be destroyed after being created.