Reallocate MPQD matrix after steady state solve
Standard PETSc reset routine allocates memory based on number of
non-zero entries before reset. As the steady state and transient
solves have different numbers of non-zeros, this can lead to a
lot of reallocating when proceeding from a steady state to
transient solve unless a conservatively large chunk of memory
is reserved before building the transient system, which is what
now happens.